Moved some classes to proper namespaces
